quote-api/requirements.txt
flask
flask-smorest
python-dotenv
psycopg2
flask-sqlalchemy
gunicorn